+++ examples/todo/src/1_schema.ts
@@ -11,7 +11,7 @@ import { Account, CoList, CoMap, Profile, co } from "jazz-tools";
  **/
 
 /** An individual task which collaborators can tick or rename */
-export class Task extends CoMap<Task> {
+export class Task extends CoMap {
     done = co.boolean;
     text = co.string;
 }
@@ -19,7 +19,7 @@ export class Task extends CoMap<Task> {
 export class ListOfTasks extends CoList.Of(co.ref(Task)) {}
 
 /** Our top level object: a project with a title, referencing a list of tasks */
-export class TodoProject extends CoMap<TodoProject> {
+export class TodoProject extends CoMap {
     title = co.string;
     tasks = co.ref(ListOfTasks);
 }
@@ -28,11 +28,11 @@ export class ListOfProjects extends CoList.Of(co.ref(TodoProject)) {}
 
 /** The account root is an app-specific per-user private `CoMap`
  *  where you can store top-level objects for that user */
-export class TodoAccountRoot extends CoMap<TodoAccountRoot> {
+export class TodoAccountRoot extends CoMap {
     projects = co.ref(ListOfProjects);
 }
 
-export class TodoAccount extends Account<TodoAccount> {
+export class TodoAccount extends Account {
     profile = co.ref(Profile);
     root = co.ref(TodoAccountRoot);
 
@@ -42,14 +42,14 @@ export class TodoAccount extends Account<TodoAccount> {
     migrate(creationProps?: { name: string }) {
         super.migrate(creationProps);
         if (!this._refs.root) {
-            this.root = new TodoAccountRoot(
+            this.root = TodoAccountRoot.create(
                 {
-                    projects: new ListOfProjects([], { owner: this }),
+                    projects: ListOfProjects.create([], { owner: this }),
                 },
                 { owner: this }
             );
         }
-    };
+    }
 }
 
 /** Walkthrough: Continue with ./2_main.tsx */